automatically generate typescript definitions
