+++ types/jsdoc-legacy.conf.json
@@ -1,0 +1,17 @@
+{
+    "source": {
+        "include": [
+            "bundles",
+            "packages"
+        ],
+        "excludePattern": "(node_modules|lib|dist|test|deprecated)"
+    },
+    "plugins": [
+        "@pixi/jsdoc-template/plugins/es6-fix"
+    ],
+    "opts": {
+        "recurse": true,
+        "destination": "./dist/types/pixi.js-legacy",
+        "template": "./node_modules/tsd-jsdoc/dist"
+    }
+}
