+++ types/jsdoc.conf.json
@@ -1,0 +1,17 @@
+{
+    "source": {
+        "include": [
+            "bundles",
+            "packages"
+        ],
+        "excludePattern": "(node_modules|lib|dist|test|deprecated|canvas)"
+    },
+    "plugins": [
+        "@pixi/jsdoc-template/plugins/es6-fix"
+    ],
+    "opts": {
+        "recurse": true,
+        "destination": "./dist/types/pixi.js",
+        "template": "./node_modules/tsd-jsdoc/dist"
+    }
+}
